Externalize Profiles
Previously all of the profile.d scripts that would be contributed by the
buildpack were stored as strings in Go code.  This was problematic as none of
the normal tooling and formatting that IDEs provide could be used to ensure
that the scripts themselves were of high quality.

This change moves all of those scripts to standalone files and uses statik to
embed them within the Go code.

Signed-off-by: Ben Hale <bhale@pivotal.io>